requirements.txt
matplotlib
numpy
pandas
paramiko
pynvml
setproctitle
pyyaml